+++ OUDS/Core/Themes/OrangeInverse/Tests/OrangeInverseThemeTests.swift
@@ -11,18 +11,26 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
+@testable import OUDS
 @testable import OUDSThemesOrangeInverse
 import Testing
 
 // swiftlint:disable force_cast
 
-/// Check if the themes ahs the default tokens providers with the expected types.
+/// Check if the theme has the default configuration and tokens providers with the expected types.
 struct OrangeInverseThemeTests {
 
     private var theme: OrangeInverseTheme
 
     init() {
         theme = OrangeInverseTheme()
+    }
+
+    // MARK: - Tests - Theme tuning
+
+    @Test func defaultTuningOfOrangeInverseTheme() throws {
+        let tuning = theme.tuning
+        #expect(tuning == Tuning.default)
     }
 
     // MARK: - Tests - Semantic tokens providers