+++ Package.swift
@@ -77,9 +77,9 @@ let package = Package(
         // Apple Swift tool to build documentation
         .package(url: "https://github.com/swiftlang/swift-docc-plugin", exact: "1.4.5"),
         // Linter for Swift code
-        .package(url: "https://github.com/SimplyDanny/SwiftLintPlugins", exact: "0.60.0"),
+        .package(url: "https://github.com/SimplyDanny/SwiftLintPlugins", exact: "0.61.0"),
         // Formatter for Swift code
-        .package(url: "https://github.com/nicklockwood/SwiftFormat", from: "0.57.2"),
+        .package(url: "https://github.com/nicklockwood/SwiftFormat", from: "0.58.0"),
     ],
 
     // MARK: - Targets